refactor(studio): extract QuickStart into its own module
Address three reviewer points on the eng-658 PR:

1. Extract `QuickStart` + `buildQuickStartSample` (and the constants
   they depend on) out of `Endpoints.tsx` into a new
   `pages/QuickStart.tsx`. `Endpoints.tsx` was already large; folding
   the Quick Start UI in compounded that and made unrelated test
   breakages (e.g. anything in the deployment list / detail tree) more
   likely to cascade. The detail page now imports a single
   `<QuickStart>` component.

2. Defer the SDK base-URL derivation to the Python / JavaScript
   branches. The cURL branch hits the per-operation URL directly and
   doesn't need `/v1` derivation, so the `URL`-parse + `try/catch`
   was paid for unnecessarily on every keystroke.

3. Rename the test fixture constant `URL` → `ENDPOINT_URL`. The
   previous name shadowed the global `URL` constructor, which the
   production module itself uses for SDK base-URL derivation; a
   future test that wanted to assert behaviour by parsing the
   sample with `new URL(...)` would have hit the shadow. The test
   file is also renamed `Endpoints.quickstart.test.ts` →
   `QuickStart.test.ts` so it co-locates with the module under test.
